requirements.txt
pandas
numpy
nltk
language-tool-python
pyspellchecker
install-jdk